refactor: Migrated nativeContent.ts
- Removed namespaces
- Fixed imports
- @TODO: Documentation
- @TODO: Check the code and move functionality to the right place if
  needed
